backend/pyproject.toml
[project]
name = "trace911-backend"
version = "0.1.0"
description = "trace911 — emergency call intelligence backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "python-multipart>=0.0.9",
    "numpy>=1.26",
    # DB
    "sqlalchemy[asyncio]>=2.0.30",
    "aiosqlite>=0.20",
    # Audio + AI pipeline (wire up as you implement services/)
    "livekit-agents>=0.11",
    "livekit-plugins-ai-coustics>=0.1",
    # STT
    "httpx>=0.27",
    # LLM extraction (pick one when you wire it)
    # "anthropic>=0.39",
    # "openai>=1.40",
]

[project.optional-dependencies]
dev = ["ruff>=0.6", "pytest>=8", "pytest-asyncio>=0.23"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
